requirements.txt
torch >= 2.1.0
datasets >= 1.1.3
langchain >= 0.1.12
langchain-openai >= 0.0.8
python-dotenv >= 0.0.5
sentence-transformers >= 0.4.1
faiss-cpu >= 1.6.3
hydra-core >= 1.3
omegaconf >= 2.3.0
